# MQE: eliminate loops in `BucketedPool`