--- a/pkg/streamingpromql/types/ring_buffer_test.go
+++ b/pkg/streamingpromql/types/ring_buffer_test.go
@@ -19,7 +19,7 @@ type ringBuffer[T any] interface {
 	DiscardPointsAtOrBefore(t int64)
 	Append(p T) error
 	Reset()
-	Use(s []T)
+	Use(s []T) error
 	Release()
 	ViewUntilSearchingForwardsForTesting(maxT int64) ringBufferView[T]
 	ViewUntilSearchingBackwardsForTesting(maxT int64) ringBufferView[T]
@@ -120,7 +120,8 @@ func testRingBuffer[T any](t *testing.T, buf ringBuffer[T], points []T) {
 
 	pointsWithPowerOfTwoCapacity := make([]T, 0, 16) // Use must be passed a slice with a capacity that is equal to a power of 2.
 	pointsWithPowerOfTwoCapacity = append(pointsWithPowerOfTwoCapacity, points...)
-	buf.Use(pointsWithPowerOfTwoCapacity)
+	err := buf.Use(pointsWithPowerOfTwoCapacity)
+	require.NoError(t, err)
 	shouldHavePoints(t, buf, points...)
 
 	buf.DiscardPointsAtOrBefore(4)
@@ -131,7 +132,8 @@ func testRingBuffer[T any](t *testing.T, buf ringBuffer[T], points []T) {
 
 	subsliceWithPowerOfTwoCapacity := make([]T, 0, 8) // Use must be passed a slice with a capacity that is equal to a power of 2.
 	subsliceWithPowerOfTwoCapacity = append(subsliceWithPowerOfTwoCapacity, points[4:]...)
-	buf.Use(subsliceWithPowerOfTwoCapacity)
+	err = buf.Use(subsliceWithPowerOfTwoCapacity)
+	require.NoError(t, err)
 	shouldHavePoints(t, buf, points[4:]...)
 }