#### What this PR does

This PR eliminates the use of loops in `BucketedPool`. This gives a very mild speedup (<1%) to queries with many steps and many series.

It also removes the panics in both ring buffer's `Append` and `Use` methods. It's possible that the conditions that previously would have caused a panic could happen in the real world if a subquery selected enough points that it exceeded the maximum size of a slice returned from the `FPoint` or `HPoint` pool, so it seems safer to return an error here rather than crashing the process.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [n/a] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.