@jhesketh 2024-12-17
This introduced a bug fixed in https://github.com/grafana/mimir/pull/10261